pyproject.toml
[project]
name = "image-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.1.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
tracking = [
    "neptune>=1.14.0",
]
plotting = [
    "matplotlib>=3.10.8",
]
